[ADD]RandomGenerator.swift
