+++ drawingApp/drawingApp/Extensions/Utils.swift
@@ -33,12 +33,7 @@ extension Color {
         return String(format:"#%06x", rgb)
     }
         
-   static func getRandomColor() -> Color {
-        let red = Double.random(in: 0..<255)
-        let green = Double.random(in: 0..<255)
-        let blue = Double.random(in: 0..<255)
-        return Color(r: red, g: green, b: blue)
-    }
+  
     
     
 }
